+++ src/runez/__init__.py
@@ -14,7 +14,7 @@ It's recommended to set DRYRUN only once at the start of your run via: runez.log
 """
 
 from runez import heartbeat, program, serialize
-from runez.base import decode, prop, Slotted
+from runez.base import decode, prop, Slotted, UNSET
 from runez.context import CaptureOutput, CurrentFolder, TempFolder, TrackedOutput, verify_abort
 from runez.convert import Anchored, flattened, formatted, quoted, represented_args, resolved_path, short, shortened, to_int
 from runez.file import copy, delete, first_line, get_conf, get_lines, move, symlink, touch, write
@@ -28,7 +28,7 @@ from runez.system import abort, get_timezone, get_version, set_dryrun
 __all__ = [
     "DRYRUN",
     "heartbeat", "logsetup", "program", "serialize",
-    "decode", "prop", "Slotted",
+    "decode", "prop", "Slotted", "UNSET",
     "CaptureOutput", "CurrentFolder", "TempFolder", "TrackedOutput", "verify_abort",
     "Anchored", "flattened", "formatted", "quoted", "represented_args", "resolved_path", "short", "shortened", "to_int",
     "copy", "delete", "first_line", "get_conf", "get_lines", "move", "symlink", "touch", "write",